+++ presto-main-base/src/main/java/com/facebook/presto/operator/aggregation/state/HyperLogLogState.java
@@ -27,5 +27,5 @@ public interface HyperLogLogState
 
     void setHyperLogLog(HyperLogLog value);
 
-    void addMemoryUsage(int value);
+    void addMemoryUsage(long value);
 }
